src/dress_synthesizer.py
import google.generativeai as genai
from PIL import Image
import io
import base64
import time
import streamlit as st

class DressSynthesizer:
    """Handles AI-powered dress try-on synthesis using Google Gemini"""
    
    def __init__(self, model):
        self.model = model
        self.generation_config = {
            "temperature": 0.7,
            "top_p": 0.8,
            "top_k": 40,
            "max_output_tokens": 1000,
        }
    
    def generate_try_on(self, human_image, dress_image):
        """
        Generate virtual try-on using Gemini's vision capabilities
        """
        try:
            # Create the prompt for virtual try-on
            prompt = self._create_try_on_prompt()
            
            # Convert images to format suitable for Gemini
            human_image_data = self._prepare_image_for_gemini(human_image)
            dress_image_data = self._prepare_image_for_gemini(dress_image)
            
            # Since Gemini doesn't directly generate images, we'll use a creative approach
            # We'll describe the try-on scenario and then use that to guide image generation
            description_prompt = f"""
            Analyze these two images:
            1. A person in the first image
            2. A dress/clothing item in the second image
            
            Provide a detailed description of how the person would look wearing the dress.
            Include specific details about:
            - How the dress would fit on the person's body
            - The color combinations and how they complement the person
            - The style and how it matches the person's appearance
            - The overall look and aesthetic
            
            Be very detailed and specific about the visual result.
            """
            
            # Generate description using Gemini
            response = self.model.generate_content([
                description_prompt,
                human_image_data,
                dress_image_data
            ])
            
            # For now, we'll create a composite image as a placeholder
            # In a production environment, you'd integrate with an image generation API
            result_image = self._create_composite_result(human_image, dress_image, response.text)
            
            return result_image
            
        except Exception as e:
            st.error(f"Error in AI processing: {str(e)}")
            # Return a fallback composite image
            return self._create_simple_composite(human_image, dress_image)
    
    def _create_try_on_prompt(self):
        """Create a detailed prompt for virtual try-on"""
        return """
        You are an expert fashion stylist and virtual try-on specialist. 
        I will provide you with two images:
        1. A photo of a person
        2. A photo of a dress or clothing item
        
        Please analyze both images and provide a detailed description of how 
        the person would look wearing the dress. Consider:
        - Body shape and how the dress would fit
        - Color coordination
        - Style compatibility
        - Overall aesthetic appeal
        - Any adjustments that might be needed
        
        Be specific and detailed in your description.
        """
    
    def _prepare_image_for_gemini(self, image):
        """Convert PIL Image to format suitable for Gemini API"""
        # Convert PIL Image to bytes
        img_byte_arr = io.BytesIO()
        image.save(img_byte_arr, format='PNG')
        img_byte_arr = img_byte_arr.getvalue()
        
        # Create the image data structure for Gemini
        return {
            "mime_type": "image/png",
            "data": img_byte_arr
        }
    
    def _create_composite_result(self, human_image, dress_image, ai_description):
        """
        Create a composite result image with AI analysis
        This is a simplified version - in production, you'd use more sophisticated image generation
        """
        # Create a canvas for the result
        canvas_width = max(human_image.width, 600)
        canvas_height = human_image.height + 200
        
        # Create new image with white background
        result = Image.new('RGB', (canvas_width, canvas_height), (255, 255, 255))
        
        # Paste the human image
        human_x = (canvas_width - human_image.width) // 2
        result.paste(human_image, (human_x, 0))
        
        # Add dress image as overlay (simplified approach)
        dress_resized = dress_image.resize((150, 200), Image.Resampling.LANCZOS)
        overlay_x = canvas_width - dress_resized.width - 20
        overlay_y = 20
        
        # Create a semi-transparent overlay effect
        overlay = Image.new('RGBA', result.size, (0, 0, 0, 0))
        overlay.paste(dress_resized, (overlay_x, overlay_y))
        
        # Convert result to RGBA for blending
        result = result.convert('RGBA')
        result = Image.alpha_composite(result, overlay)
        result = result.convert('RGB')
        
        return result
    
    def _create_simple_composite(self, human_image, dress_image):
        """Create a simple side-by-side composite as fallback"""
        # Calculate dimensions
        max_height = max(human_image.height, dress_image.height)
        total_width = human_image.width + dress_image.width + 40
        
        # Create canvas
        composite = Image.new('RGB', (total_width, max_height + 100), (240, 240, 240))
        
        # Paste human image
        human_y = (max_height - human_image.height) // 2
        composite.paste(human_image, (20, human_y))
        
        # Paste dress image
        dress_y = (max_height - dress_image.height) // 2
        dress_x = human_image.width + 40
        composite.paste(dress_image, (dress_x, dress_y))
        
        return composite
    
    def analyze_compatibility(self, human_image, dress_image):
        """Analyze style compatibility between person and dress"""
        try:
            prompt = """
            Analyze the compatibility between this person and this dress/clothing item.
            Consider:
            1. Color coordination
            2. Style compatibility
            3. Fit predictions
            4. Overall aesthetic appeal
            5. Suggestions for improvement
            
            Provide a detailed analysis with a compatibility score (1-10).
            """
            
            human_data = self._prepare_image_for_gemini(human_image)
            dress_data = self._prepare_image_for_gemini(dress_image)
            
            response = self.model.generate_content([
                prompt,
                human_data,
                dress_data
            ])
            
            return response.text
            
        except Exception as e:
            return f"Analysis unavailable: {str(e)}"
    
    def get_styling_suggestions(self, human_image, dress_image):
        """Get AI-powered styling suggestions"""
        try:
            prompt = """
            Based on this person and dress combination, provide styling suggestions:
            1. Recommended accessories
            2. Shoe suggestions
            3. Hair and makeup recommendations
            4. Color coordination tips
            5. Occasion suitability
            
            Be specific and practical in your suggestions.
            """
            
            human_data = self._prepare_image_for_gemini(human_image)
            dress_data = self._prepare_image_for_gemini(dress_image)
            
            response = self.model.generate_content([
                prompt,
                human_data,
                dress_data
            ])
            
            return response.text
            
        except Exception as e:
            return f"Suggestions unavailable: {str(e)}"